remove extracted funxtio
